+++ types/BillingPortal/Configurations.d.ts
@@ -4,7 +4,7 @@ declare module 'stripe' {
   namespace Stripe {
     namespace BillingPortal {
       /**
-       * A portal configuration describes the functionality and behavior of a portal session.
+       * A portal configuration describes the functionality and behavior you embed in a portal session. Related guide: [Configure the customer portal](https://docs.stripe.com/customer-management/configure-portal).
        */
       interface Configuration {
         /**