+++ types/Charges.d.ts
@@ -1538,7 +1538,7 @@ declare module 'stripe' {
         }
 
         namespace Crypto {
-          type Network = 'base' | 'ethereum' | 'polygon';
+          type Network = 'base' | 'ethereum' | 'polygon' | 'solana';
 
           type TokenCurrency = 'usdc' | 'usdg' | 'usdp';
         }
@@ -2359,7 +2359,12 @@ declare module 'stripe' {
 
         interface Qris {}
 
-        interface Rechnung {}
+        interface Rechnung {
+          /**
+           * Payment portal URL.
+           */
+          payment_portal_url: string | null;
+        }
 
         interface RevolutPay {
           funding?: RevolutPay.Funding;