+++ types/ConfirmationTokens.d.ts
@@ -1645,6 +1645,7 @@ declare module 'stripe' {
           | 'card_present'
           | 'cashapp'
           | 'crypto'
+          | 'custom'
           | 'customer_balance'
           | 'eps'
           | 'fpx'