+++ types/V2/Core/AccountsResource.d.ts
@@ -121,6 +121,7 @@ declare module 'stripe' {
                 type LocationSource =
                   | 'identity_address'
                   | 'ip_address'
+                  | 'payment_method'
                   | 'shipping_address';
               }
 
@@ -1056,10 +1057,22 @@ declare module 'stripe' {
                    * Can hold storage-type funds on Stripe in GBP.
                    */
                   gbp?: HoldsCurrencies.Gbp;
+
+                  /**
+                   * Can hold storage-type funds on Stripe in USD.
+                   */
+                  usd?: HoldsCurrencies.Usd;
                 }
 
                 namespace HoldsCurrencies {
                   interface Gbp {
+                    /**
+                     * To request a new Capability for an account, pass true. There can be a delay before the requested Capability becomes active.
+                     */
+                    requested: boolean;
+                  }
+
+                  interface Usd {
                     /**
                      * To request a new Capability for an account, pass true. There can be a delay before the requested Capability becomes active.
                      */
@@ -1295,7 +1308,11 @@ declare module 'stripe' {
             }
 
             namespace Responsibilities {
-              type FeesCollector = 'application' | 'stripe';
+              type FeesCollector =
+                | 'application'
+                | 'application_custom'
+                | 'application_express'
+                | 'stripe';
 
               type LossesCollector = 'application' | 'stripe';
             }
@@ -1303,7 +1320,7 @@ declare module 'stripe' {
 
           interface Identity {
             /**
-             * Attestations from the identity's key people, e.g. owners, executives, directors.
+             * Attestations from the identity's key people, e.g. owners, executives, directors, representatives.
              */
             attestations?: Identity.Attestations;
 
@@ -1344,6 +1361,11 @@ declare module 'stripe' {
                * Attestation that all Persons with a specific Relationship value have been provided.
                */
               persons_provided?: Attestations.PersonsProvided;
+
+              /**
+               * This hash is used to attest that the representative is authorized to act as the representative of their legal entity.
+               */
+              representative_declaration?: Attestations.RepresentativeDeclaration;
 
               /**
                * Attestations of accepted terms of service agreements.
@@ -1412,6 +1434,23 @@ declare module 'stripe' {
                 type OwnershipExemptionReason =
                   | 'qualified_entity_exceeds_ownership_threshold'
                   | 'qualifies_as_financial_institution';
+              }
+
+              interface RepresentativeDeclaration {
+                /**
+                 * The time marking when the representative attestation was made. Represented as a RFC 3339 date & time UTC value in millisecond precision, for example: 2022-09-18T13:22:18.123Z.
+                 */
+                date?: string;
+
+                /**
+                 * The IP address from which the representative attestation was made.
+                 */
+                ip?: string;
+
+                /**
+                 * The user agent of the browser from which the representative attestation was made.
+                 */
+                user_agent?: string;
               }
 
               interface TermsOfService {
@@ -2702,6 +2741,7 @@ declare module 'stripe' {
                 type LocationSource =
                   | 'identity_address'
                   | 'ip_address'
+                  | 'payment_method'
                   | 'shipping_address';
 
                 type ValidateLocation = 'auto' | 'deferred' | 'immediately';
@@ -3625,10 +3665,22 @@ declare module 'stripe' {
                    * Can hold storage-type funds on Stripe in GBP.
                    */
                   gbp?: HoldsCurrencies.Gbp;
+
+                  /**
+                   * Can hold storage-type funds on Stripe in USD.
+                   */
+                  usd?: HoldsCurrencies.Usd;
                 }
 
                 namespace HoldsCurrencies {
                   interface Gbp {
+                    /**
+                     * To request a new Capability for an account, pass true. There can be a delay before the requested Capability becomes active.
+                     */
+                    requested?: boolean;
+                  }
+
+                  interface Usd {
                     /**
                      * To request a new Capability for an account, pass true. There can be a delay before the requested Capability becomes active.
                      */
@@ -3864,7 +3916,11 @@ declare module 'stripe' {
             }
 
             namespace Responsibilities {
-              type FeesCollector = 'application' | 'stripe';
+              type FeesCollector =
+                | 'application'
+                | 'application_custom'
+                | 'application_express'
+                | 'stripe';
 
               type LossesCollector = 'application' | 'stripe';
             }
@@ -3872,7 +3928,7 @@ declare module 'stripe' {
 
           interface Identity {
             /**
-             * Attestations from the identity's key people, e.g. owners, executives, directors.
+             * Attestations from the identity's key people, e.g. owners, executives, directors, representatives.
              */
             attestations?: Identity.Attestations;
 
@@ -3913,6 +3969,11 @@ declare module 'stripe' {
                * Attestation that all Persons with a specific Relationship value have been provided.
                */
               persons_provided?: Attestations.PersonsProvided;
+
+              /**
+               * This hash is used to attest that the representative is authorized to act as the representative of their legal entity.
+               */
+              representative_declaration?: Attestations.RepresentativeDeclaration;
 
               /**
                * Attestations of accepted terms of service agreements.
@@ -3981,6 +4042,23 @@ declare module 'stripe' {
                 type OwnershipExemptionReason =
                   | 'qualified_entity_exceeds_ownership_threshold'
                   | 'qualifies_as_financial_institution';
+              }
+
+              interface RepresentativeDeclaration {
+                /**
+                 * The time marking when the representative attestation was made. Represented as a RFC 3339 date & time UTC value in millisecond precision, for example: 2022-09-18T13:22:18.123Z.
+                 */
+                date?: string;
+
+                /**
+                 * The IP address from which the representative attestation was made.
+                 */
+                ip?: string;
+
+                /**
+                 * The user agent of the browser from which the representative attestation was made.
+                 */
+                user_agent?: string;
               }
 
               interface TermsOfService {
@@ -4905,6 +4983,11 @@ declare module 'stripe' {
           applied_configurations?: Array<string>;
 
           /**
+           * Filter by whether the account is closed. If omitted, returns only Accounts that are not closed.
+           */
+          closed?: boolean;
+
+          /**
            * The upper limit on the number of accounts returned by the List Account request.
            */
           limit?: number;
@@ -4981,7 +5064,7 @@ declare module 'stripe' {
           ): ApiListPromise<Stripe.V2.Core.Account>;
 
           /**
-           * Removes access to the Account and its associated resources.
+           * Removes access to the Account and its associated resources. Closed Accounts can no longer be operated on, but limited information can still be retrieved through the API in order to be able to track their history.
            */
           close(
             id: string,