+++ types/V2/Core/EventTypes.d.ts
@@ -5,7 +5,6 @@ declare module 'stripe' {
     export type Event =
       | Stripe.Events.V1BillingMeterErrorReportTriggeredEvent
       | Stripe.Events.V1BillingMeterNoMeterFoundEvent
-      | Stripe.Events.V2BillingBillSettingUpdatedEvent
       | Stripe.Events.V2CoreAccountClosedEvent
       | Stripe.Events.V2CoreAccountCreatedEvent
       | Stripe.Events.V2CoreAccountUpdatedEvent
@@ -71,7 +70,6 @@ declare module 'stripe' {
     export type EventNotification =
       | Stripe.Events.V1BillingMeterErrorReportTriggeredEventNotification
       | Stripe.Events.V1BillingMeterNoMeterFoundEventNotification
-      | Stripe.Events.V2BillingBillSettingUpdatedEventNotification
       | Stripe.Events.V2CoreAccountClosedEventNotification
       | Stripe.Events.V2CoreAccountCreatedEventNotification
       | Stripe.Events.V2CoreAccountUpdatedEventNotification
@@ -352,38 +350,6 @@ declare module 'stripe' {
             }
           }
         }
-      }
-    }
-
-    /**
-     * This event occurs when a bill setting is updated.
-     */
-    export interface V2BillingBillSettingUpdatedEvent
-      extends V2.Core.EventBase {
-      type: 'v2.billing.bill_setting.updated';
-      // Retrieves data specific to this event.
-      data: V2BillingBillSettingUpdatedEvent.Data;
-      // Object containing the reference to API resource relevant to the event.
-      related_object: V2.Core.Events.RelatedObject;
-      // Retrieves the object associated with the event.
-      fetchRelatedObject(): Promise<V2.Billing.BillSetting>;
-    }
-    export interface V2BillingBillSettingUpdatedEventNotification
-      extends V2.Core.EventNotificationBase {
-      type: 'v2.billing.bill_setting.updated';
-      // Object containing the reference to API resource relevant to the event.
-      related_object: V2.Core.Events.RelatedObject;
-      // Retrieves the object associated with the event.
-      fetchRelatedObject(): Promise<V2.Billing.BillSetting>;
-      fetchEvent(): Promise<V2BillingBillSettingUpdatedEvent>;
-    }
-
-    namespace V2BillingBillSettingUpdatedEvent {
-      export interface Data {
-        /**
-         * Timestamp of when the object was updated.
-         */
-        updated: string;
       }
     }
 
@@ -1702,6 +1668,8 @@ declare module 'stripe' {
     export interface V2MoneyManagementTransactionCreatedEvent
       extends V2.Core.EventBase {
       type: 'v2.money_management.transaction.created';
+      // Retrieves data specific to this event.
+      data: V2MoneyManagementTransactionCreatedEvent.Data;
       // Object containing the reference to API resource relevant to the event.
       related_object: V2.Core.Events.RelatedObject;
       // Retrieves the object associated with the event.
@@ -1715,6 +1683,15 @@ declare module 'stripe' {
       // Retrieves the object associated with the event.
       fetchRelatedObject(): Promise<V2.MoneyManagement.Transaction>;
       fetchEvent(): Promise<V2MoneyManagementTransactionCreatedEvent>;
+    }
+
+    namespace V2MoneyManagementTransactionCreatedEvent {
+      export interface Data {
+        /**
+         * Id of the v1 Transaction corresponding to this Transaction.
+         */
+        v1_id?: string;
+      }
     }
 
     /**