+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 ## 19.2.0-beta.1 - 2025-10-03
 * Contains bug fixes and improvements from [v19.1.0](https://github.com/stripe/stripe-node/blob/master/CHANGELOG.md#1910---2025-10-03).
 
+## 19.1.0 - 2025-10-03
+* [#2453](https://github.com/stripe/stripe-node/pull/2453) add missing fetchEvent type for UnknownEventNotification
+  
+  - Add missing `fetchEvent()` declaration to the `Stripe.Events.UnknownEventNotification` interface
+  - Tweak `Stripe.Events.fetchRelatedObject` so that it's always defined and returns `null` if there's no `related_object`. This fixes the situation where the `UnknownEventNotification` says that `fetchRelatedObject()` is defined, but calling it throws an error.
+* [#2447](https://github.com/stripe/stripe-node/pull/2447) Update param in deprecation docs link
+* [#2444](https://github.com/stripe/stripe-node/pull/2444) Update CHANGELOG.md to point to right API version
+
 ## 19.1.0-beta.1 - 2025-09-30
 This release changes the pinned API version to `2025-09-30.preview`. It is built on top of SDK version 19.0.0 which contains breaking changes. Please review the [changelog for 19.0.0](https://github.com/stripe/stripe-go/blob/master/CHANGELOG.md#1900---2025-09-30) if upgrading from older SDK versions.